This is probably the version right before the one that works.
The API work between CxFlow and CxOD is complete. Scans be created and
business unit/project management is working.

There still needs to be a little work with preset selection. The issues
still a couple of small tweaks to the final presentation and the code to
wait until scans are complete still needs tested.