The namespace now maps to CxOD application and project name now maps to CxOD project.
The CxOD API will return business units and applications but it will not return ones that are the children of others. This resulted in situations where BU's or applications with the same names but under different relative paths would sometimes colid. This version of the fixed that by downloading the entire navigation three and parsing it real time to discover which BU's and applications own each other.